Ready?
I *think* I’m ready to start the creation of nodes and trees.  I have
to decide if I want to search forward or backward.  The original
literature used a backward search, possibly with some kind of
constraints to enforce ordering.  If you used forward search, it seems
like you could NOT execute an action if its preconditions had been met
and thus force preconditions to be satisfied.  This would prune a lot
of possible paths making the tree more narrow.  It also guarantees the
ordering is enforced because you can’t execute the action without the
preconditions being satisfied.  If it works as expected…