Fixed bug in looping.
It works for the basic forward chaining solution, giving this result:

Actions:
  -  ('Agent', 'Activate Door', 'Access Door Activator 1')
  -  ('Agent', 'Pick Up Object', 'Red Card')
  -  ('Agent', 'Go Through Door', 'Access Door')
  -  ('Agent', 'Activate RA Door', 'Red Access Door Activator 1')
  -  ('Agent', 'Go Through Door', 'Red Access Door')
  -  ('Agent', 'Activate Shuttle Gen', 'Shuttle Generator')
  -  ('Agent', 'Activate Shuttle', 'Shuttle Activator')

WHEN THE PLAYER STARTS IN room cRoom1.

But if they start in cRoom2, no solution is found.  And this makes
sense, because if the player goes out of room 2 into room 1 to pick up
the card, they have to come back into room 2, which would mean using an
action that had already been executed (‘Agent’, ‘Go Through Door’,
‘Access Door’).  It may be that instead of saying “you can only perform
this action once”, it should be “you can only perform this action once
per room”.  So if the action is present in the last N moves, where N is
the number of valid actions in the room, then it is invalid.
Otherwise, it may be on the list…