Working Better
I removed the constraint that actions could NOT be repeated and now the
ForwardPlanner works finding a solution regardless of the room the
agent starts in.  HOWEVER, the cost of this is that the number of nodes
explored is much higher when going from room 1 (4 vs. 27).  There may
be some balancing to improve the number of nodes expanded, but this is
probably going to be a tradeoff issue.

So, maybe the best solution is to try to find a solution with “no
repeated actions” turned on.  If we find it, great.  If we don’t, then
try again with it turned off and an iteration limit.  If we find it,
great, if not, we can’t find a solution.