Added Mixed Strategy, one experiment showed some interesting effects.
I have added a mixed strategy to the planner where it first tries using
the constraint that all actions are “unique” in the history.  If this
fails, which it usually does quickly, it will try without the
constraint and a limit on the number of node expansions it will try.
This yields a nice balance of results.

I also tried having the doors automatically be closed by the action of
going through them.  In a real game, the doors would close soon after
somebody opened them and would stay open if somebody was in doorway.
Having the door close really increased the number of iterations (2x)
when the agent had to go back and forth to solve the problem.

However, I don’t feel having the action do this is legitimate.   If the
agent triggers the door then goes through and needs to go back through
it, it may still be open.  If it is closed, the plan will be
invalidated and re-evaluated.  But all it takes is one hiccup in plan
and it may need to be re-evaluated, so this is not necessarily buying
us anything to have it “planned in” that the door will close behind us.