Got rid of "Has Red Access"
Fixed bug in finding objects in the room.  I forgot to clear out the
list from the previous loop, so the old list was in there and the Red
Access Key was “in the room” even though it was being carried.

Using the Red Access Door Activator requires you to have picked up the
Red Access Key.  There are several ways to make this happen:
1. Have the “Has Red Access” flag set to true when you pick up the Key.
 When you drop it, you would have to set it to False.  More than one
key => Complication.  Got Red Access for some other reason =>
Complication.  You could handle the last one by having different kinds
of Red Access, but that is also complicated.
2. Have the Red Access Activator check if you have the proper “token”.
Like a mutex, you either have it or you do not have it.  Have
multiples, not a problem.  But requires a procedural check instead of a
logical one.
3. When you pick up the key, it adds the action to use the activator.
But this leads to the same complications when dropping the key or when
you have multiple keys.

For now, I am going to use the “Inventory Controlled Mechanism”.

It is also a good example of how to use a procedural mechanism to prune
paths in the tree, which forces serial execution of operations.

I may try to do the same thing for the launch sequence…currently the
“is powered” is a precondition of doing the “is activated” for the
shuttle launch.  This is currently implemented as a precondition of the
action.  I am not sure there is much to be gained by making it
procedural.

I still feel “twitchy” about using the procedural preconditions for
backwards chaining.  Going forward, it prunes paths of exploration.
Going backward…preconditions that are not satisfied get added to the
list of goals…it seems like you could not add a procedural check to the
list of goals.  Well, you could, but it seems like you have to pass
around the function pointer at that point, along with the arguments
(closure).  This seems like a headache and I am not sure what you would
gain from it…is backwards any better than forwards?