Added iterations/action statistic to the output.
We can look at this or the number of nodes expanded or the total depth
on the open list as a measure of how “narrow” the tree was made in the
search.  So far, it looks to be from 1.0 (really good) to 1.5 (is this
good?).

Instead of looking at backward chaining, I think the next big
modification will be carrying the difference set into the nodes instead
of the modified world state.  At each node, the world state would NOT
be modified, but the difference set would.  Queries would first try the
difference set and then try the world state.  Since the difference set
is MUCH smaller than the world state, this will be a huge savings in a
real game.