Further progress on processing.
Got all the objects parsed out and an ID assigned to each one, along
with all the cell indexes that are part of it (e.g. all the cells that
are in the computer core, etc.).  This required a small fill-like
search algorithm (DFS or BFS?).  Some small kinks to work out, but
seems to work as expected.