Further segmenting data.
Putting “object” into “cell info” etc.  There are three main sets of
data:  cellInfo, roomInfo, and gameObjectInfo.  All these are cross
referential through the subjectID, cellID, and room name.  Should this
be referential through a Room ID?  That is to say, instead of making
the room name “unique”, should it just be an ID like the game object?
This does track and would get us out of a problem with duplicate room
names…